backend/requirements.txt
beautifulsoup4
requests
python-dotenv
supabase
datetime
regex
tqdm
openai